64bit compile fixes
git-svn-id: svn://svn.icculus.org/gtkradiant/GtkRadiant/trunk@106 8a3a26a2-13c4-0310-b231-cf6edde360e5

+++ radiant/eclass_doom3.cpp
@@ -668,7 +668,7 @@ bool EntityClassDoom3_parse(TextInputStream& inputStream, const char* filename)
     CopiedString tmp(blockType);
     if(!EntityClassDoom3_parseBlock(tokeniser, tmp.c_str()))
     {
-      globalErrorStream() << GlobalFileSystem().findFile(filename) << filename << ":" << tokeniser.getLine() << ": " << tmp.c_str() << " parse failed, skipping rest of file\n";
+      globalErrorStream() << GlobalFileSystem().findFile(filename) << filename << ":" << (unsigned int)tokeniser.getLine() << ": " << tmp.c_str() << " parse failed, skipping rest of file\n";
       return false;
     }
   }
